MMPI Baseline source code and TeraGrid Measurements
This contains a simple MPI program (mmpi.c) that measures communication
times between cores on a computer cluster and output a set of
measurements in a text file. There are associated python scripts that
will process the measurement files and matlab scripts to plot the
results as a type of surface contour. Measurements are shown as
core-to-core communications time and variability in communication
measurements. The primary output data product produced by the software
is a visual representation of communications patterns on a cluster. The
software was developed during the NSF TeraGrid era. Included in this
repository are several measurements taken of TeraGrid systems all of
which have been decommissioned.