Create .gitattributes
Add git attributes so that program type changes on this public repo. The main program here is in C, and it requires MPI libraries. The matlab scripts in the repo are primarily post-processing plotting scripts, and are not essential to the main purpose of this software, which was to measure MPI delays on different types of supercomputers.